project/requirements.txt
fastapi[standard]
uvicorn
psycopg2-binary
python-dotenv
requests
pydantic[email]
python-jose[cryptography]
bcrypt
python-multipart
pytz
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
msal
